close #800

+++ lib/src/pages/download/grid/archive/archive_grid_download_page_logic.dart
@@ -52,6 +52,10 @@ class ArchiveGridDownloadPageLogic extends GetxController
 
   @override
   Future<void> handleRemoveItem(ArchiveDownloadedData archive) async {
+    bool confirmed = await confirmDestructiveAction(title: 'delete'.tr + '?');
+    if (!confirmed) {
+      return;
+    }
     await archiveDownloadService.deleteArchive(archive.gid).then((_) => super.handleRemoveItem(archive));
     updateGlobalGalleryStatus();
   }
